Move visualization functions to new module file visualization_utils.py
