+++ src/viqa/__init__.py
@@ -6,7 +6,7 @@ The package provides a set of metrics to assess the quality of volumetric images
 It can be used for 2D and 3D images.
 """
 
-__version__ = "0.19.4"
+__version__ = "0.19.5"
 __author__ = "Lukas Behammer"
 __all__ = ["FSIM", "GSM", "MAD", "UQI", "MSSSIM", "PSNR", "PSNR", "RMSE", "SSIM",
            "VIFp", "VSI", "CNR", "SNR", "QMeasure", "load_data", "normalize_data",
